+++ core/src/commands/cmd_diff.rs
@@ -7,7 +7,7 @@ use colored::Colorize;
 use crate::{
     backend::{BackendOptions, new_backend_with_prompt},
     commands::{GlobalArgs, cleanup::CleanupHandler},
-    fs::tree::{NodeDiff, NodeDiffStreamer, SerializedNodeStream},
+    fs::tree::{NodeDiff, NodeDiffStream, SerializedNodeStream},
     mapache::{ContentIdType, defaults::SHORT_SNAPSHOT_ID_LEN},
     repository::{
         repo::{RepoConfig, Repository},
@@ -70,21 +70,21 @@ pub fn run(global_args: &GlobalArgs, args: &CmdArgs) -> Result<()> {
     let source_snapshot = repo.load_snapshot(&source_id, None)?;
     let target_snapshot = repo.load_snapshot(&target_id, None)?;
 
-    let source_node_streamer = SerializedNodeStream::new(
+    let source_node_stream = SerializedNodeStream::new(
         repo.clone(),
         Some(source_snapshot.tree),
         PathBuf::new(),
         args.include.clone(),
         args.exclude.clone(),
     )?;
-    let target_node_streamer = SerializedNodeStream::new(
+    let target_node_stream = SerializedNodeStream::new(
         repo.clone(),
         Some(target_snapshot.tree),
         PathBuf::new(),
         args.include.clone(),
         args.exclude.clone(),
     )?;
-    let diff_streamer = NodeDiffStreamer::new(source_node_streamer, target_node_streamer);
+    let diff_stream = NodeDiffStream::new(source_node_stream, target_node_stream);
 
     ui::cli::log!(
         "Finding diffs {}..{}\n",
@@ -96,7 +96,7 @@ pub fn run(global_args: &GlobalArgs, args: &CmdArgs) -> Result<()> {
     );
 
     let mut counts = DiffCounts::default();
-    for (path, source, target, diff_type) in diff_streamer.flatten() {
+    for (path, source, target, diff_type) in diff_stream.flatten() {
         match &diff_type {
             NodeDiff::New => {
                 let target_node = &target